test(agent): split prepared receiver coverage
